requirements.txt
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pandas>=1.5.0
tqdm>=4.64.0
openai>=1.0.0
